% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pgpTList.R
\name{anova.pgpTList}
\alias{anova.pgpTList}
\title{Log-Likelihood Ratio for \code{pgpTList} objects}
\usage{
\method{anova}{pgpTList}(object, object1, which = c("total", "GP", "timePoisson"), ...)
}
\arguments{
\item{object, object1}{Two \code{pgpTlist} objects for with
\code{object} nested in \code{object1}.}

\item{which}{The log-likelihood to be used. With the values
\code{"GP"} and \code{"timePoisson"}, only the corresponding
part of the Poisson-GP model is used. With the value
\code{"total"} the tola log-likelihood is used.}

\item{...}{Not used yet.}
}
\value{
A matrix with the results likelihood-ratio tests, one by
    threshold.
}
\description{
Log-Likelihood Ratio for \code{pgpTList} objects
}
