% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fevList.R
\name{as.fevdList}
\alias{as.fevdList}
\title{Coerce into a \code{fevdList} object.}
\usage{
as.fevdList(object)
}
\arguments{
\item{object}{An object to coerce, typically a list object.}
}
\value{
An object with S3 class \code{"fevdList"}
}
\description{
Coerce into a \code{fevdList} object.
}
