% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rqList.R
\name{as.rqTList}
\alias{as.rqTList}
\title{Coerce into a \code{rqTList} object.}
\usage{
as.rqTList(object)
}
\arguments{
\item{object}{An object to coerce, typically a list object.}
}
\value{
An object with S3 class \code{"rqTList"}
}
\description{
Coerce into a \code{rqTList} object.
}
\note{
It is a good practice to use a \emph{named} list.
}
