% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplotStuff.R
\name{autolayer.rqTList}
\alias{autolayer.rqTList}
\title{Autolayer Method for the Class \code{"rqTList"}.}
\usage{
\method{autolayer}{rqTList}(object, lastFullYear = TRUE, ...)
}
\arguments{
\item{object}{The \code{rqTList} object.}

\item{lastFullYear}{Logical. If \code{TRUE} the prediction will be
computed only on the last full year of the data used in the
fit. See \code{\link{predict.rqTList}}.}

\item{...}{Arguments passed to \code{\link{geom_line}} such as
\code{linetype}, \code{size}.}
}
\description{
Create a new layer on an existing ggplot object showing the
quantile regression curves i.e. the threshold curves against the
date. This is useful e.g. to show the thresholds on the same
ggplot as the timeseries used fo fit \code{object}.
}
\seealso{
\code{\link{autoplot.rqTList}}.
}
