% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplotStuff.R
\name{autoplot.dailyMet}
\alias{autoplot.dailyMet}
\title{Autoplot a daily meteorological series.}
\usage{
\method{autoplot}{dailyMet}(object, group = c("decade", "year"))
}
\arguments{
\item{object}{The object to be (gg)-ploted.}

\item{group}{A character giving the grouping variable.}
}
\value{
An object inheriting from the \code{"ggplot"} class.
}
\description{
Autoplot a daily meteorological series, usually a very long period
of time, several dozens of years.
}
