% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miscUtils.R
\name{checkDesign}
\alias{checkDesign}
\title{Rough Check for a List of Designs}
\usage{
checkDesign(designList, dt, dropDup = "Cst")
}
\arguments{
\item{designList}{A list of (definitions of) designs as can be
used in \code{\link{designVars}}.}

\item{dt}{Date or time to check the calls.}

\item{dropDup}{}
}
\value{
\code{TRUE} if the check is OK.
}
\description{
Rough Check for a List of Designs
}
\examples{
dl <- list("breaks" = list(what = "NSGEV::breaksX",
            args = list(breaks = c("1970-01-01", "1990-01-01"))))
checkDesign(dl)
dl <- list("breaks" = list(what = "NSGEV::breaksX",
            foo = list(breaks = c("1970-01-01", "1990-01-01"))))
checkDesign(dl)

}
\keyword{internal}
