% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bases.R
\name{checkSinPhiNames}
\alias{checkSinPhiNames}
\title{Check a Vector of Names for Trigonometric Components with
    Given Phase}
\usage{
checkSinPhiNames(names)
}
\arguments{
\item{names}{A character vector containing names of variables,
possibly embeding trigonometric variables. Can also be a
formula from which the names are extracted.}
}
\value{
The positive integer \code{K} such that the names
    \code{"sinPhij1"}, ..., \code{"sinPhiK"}, are found in
    \code{name}. If no trigonometric term is found then the value
    \code{0} is returned. If the harmonics do not form an interval
    \code{12:K}, an error is thrown.
}
\description{
Check whether the names of trigonometric terms with given phase
are used in a character vector or formula. A name is such
trigonometric term if it is obtained by pasting \code{"sinPhij"}
with an integer giving the harmonic. We expect that the harmonics
used form an interval \code{1:K} where \code{K} is the order.
}
\examples{
nm <- c("Cst", "sinjPhi1", "sinjPhi2", "sinjPhi3")
checkSinPhiNames(nm)
## error: no 'sinjPhi1' term
try(checkSinPhiNames(c("Cst", "sinjPhi2")))

}
