% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bases.R
\name{checkTrigNames}
\alias{checkTrigNames}
\title{Check the Names for Trigonometric Components}
\usage{
checkTrigNames(name)
}
\arguments{
\item{name}{A character vector containing the names of
trigonometric variables.}
}
\value{
The positive integer \code{K} such that the names
    \code{"cosj1"}, \code{"sinj1"}, ..., \code{"cosjK"}, \code{"sinjK"}
    are found in \code{name}.
}
\description{
Check the Names for Trigonometric Components
}
\section{Caution}{
 This fonction is likely to be come an internal
    (non-exported) function.
}

\examples{
nm <- c("Cst", "sinj1", "cosj1", "sinj2", "cosj2", "sinj3", "cosj3")
checkTrigNames(nm)

}
