% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miscUtils.R
\name{clusters3}
\alias{clusters3}
\title{Compute Clusters}
\usage{
clusters3(Date, y, u)
}
\arguments{
\item{Date}{A \code{Date} object.}

\item{y}{A numeric vector representing a daily time series. Must
have the same length as \code{Date}.}

\item{u}{The threshold (vector). If \code{u} has length one, it
will be recycled to reach the length of \code{y} and
\code{Date}.}
}
\value{
A list with the results.
}
\description{
Compute clusters on the basis of a three-day period to separate
clusters.
}
\details{
The cluster are sequences of successive dates such that: (1) any
exceedance falls in a cluster, (2) If two exceedances are
separated by three consecutive dates of non-exceedance, then they
are in different clusters. A \code{NA} value is assumed to be a
non-exceedance.
}
\section{Caution}{
 A quite unreliable function, which will not work
    when there are exceedances at the ends. There exists much
    better function that can be used for declustering.
}

\examples{
data(Fort)
Fort <- within(subset(Fort, year == 1999), {
   Prec <- 2.54 * Prec;
   Date <- as.Date(sprintf("\%4d-\%02d-\%02d", year, month, day));
})
u0 <- quantile(Fort$Prec, prob = 0.97)
cl <- clusters3(Date = Fort$Date, y = Fort$Prec, u = u0)
plot(Prec ~ Date, data = Fort, type = "h", lwd = 2)
abline(h = u0, col = "SpringGreen3")
abline(v = cl$DateClust , col = "orangered")
}
