% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fevdMethods.R
\name{coSd.fevd}
\alias{coSd.fevd}
\title{Coefficients and Standard Errors}
\usage{
\method{coSd}{fevd}(object, sOnly = FALSE, ...)
}
\arguments{
\item{object}{An object with class \code{"fevd"}.}

\item{sOnly}{Logical. If \code{TRUE} only the standard deviation
(a.k.a standard errors) wille be show.}

\item{...}{Not used.}
}
\description{
Extracts and show the coefficients of a \code{fevd} object along
with their standard deviations.
}
\examples{
tau <- 0.97
library(extRemes)
data(Fort)
u0 <- quantile(Fort$Prec, prob = tau)
fit0 <- fevd(x = Prec, data = Fort, threshold = u0, type = "GP")
coSd(fit0)
 
}
\seealso{
The \code{\link[extRemes]{distill.fevd}} method from the
    \pkg{extRemes} package.
}
