% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fevList.R, R/fevdMethods.R
\name{coSd.fevdList}
\alias{coSd.fevdList}
\title{Coefficients and Standard Errors}
\usage{
\method{coSd}{fevdList}(object, sOnly = FALSE, ...)

\method{coSd}{fevdList}(object, sOnly = FALSE, ...)
}
\arguments{
\item{object}{An object with class \code{"fevd"}.}

\item{sOnly}{Logical. If \code{TRUE} only the standard deviation
(a.k.a standard errors) wille be show.}

\item{...}{Not used.}
}
\description{
Extracts and show the coefficients of a \code{fevdList} object
along with their standard deviations.

Extracts and show the coefficients of a \code{fevdList} object
along with their standard deviations.
}
\examples{
u0 <- quantile(Fort$Prec, prob = c(0.95, 0.97, 0.98))
Fits <- list()
for (i in 1:3) {
   Fits[[i]] <- fevd(x = Prec, data = Fort, threshold = u0[i], type = "GP")
}
class(Fits) <- "fevdList"
coSd(Fits)

u0 <- quantile(Fort$Prec, prob = c(0.95, 0.97, 0.98))
Fits <- list()
for (i in 1:3) {
   Fits[[i]] <- fevd(x = Prec, data = Fort, threshold = u0[i], type = "GP")
}
class(Fits) <- "fevdList"
coSd(Fits)

}
