% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mlePPMethods.R
\name{coSd.mlePP}
\alias{coSd.mlePP}
\title{Coefficients and Standard Errors}
\usage{
\method{coSd}{mlePP}(object, sOnly = FALSE, ...)
}
\arguments{
\item{object}{An object with class \code{"mlePP"} from the
\pkg{NHPoisson} package.}

\item{sOnly}{Logical. If \code{TRUE} only the standard deviation
(a.k.a standard errors) wille be show.}

\item{...}{Not used.}
}
\description{
Extracts and show the coefficients of a \code{fevd} object along
with their standard deviations.
}
