% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fevdMethods.R
\name{coSd}
\alias{coSd}
\title{Estimated Coefficients with Standard Errors}
\usage{
coSd(object, ...)
}
\arguments{
\item{object}{A fitted model.}

\item{...}{Further arguments for methods.}
}
\value{
A character vector. The method is essentially designed to
    be used to see the result in an interactive session as usually
    \code{\link[base]{summary}}.
}
\description{
Extracts the coefficients and the corresponding standard deviation
for the estimated coefficients of fitted model.
}
