% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rqMethods.R
\name{coSd.rq}
\alias{coSd.rq}
\title{Coefficients and Standard Errors}
\usage{
\method{coSd}{rq}(object, sOnly = FALSE, ...)
}
\arguments{
\item{object}{An object with class \code{"rq"}.}

\item{sOnly}{Logical. If \code{TRUE} only the standard deviation
(a.k.a standard errors) wille be show.}

\item{...}{Not used.}
}
\description{
Extracts and show the coefficients of a \code{rq} object along
with their standard deviations.
}
