% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rqTList.R
\name{coSd.rqTList}
\alias{coSd.rqTList}
\title{Coefficients and Standard Errors}
\usage{
\method{coSd}{rqTList}(object, sOnly = FALSE, ...)
}
\arguments{
\item{object}{An object with class \code{"rq"}.}

\item{sOnly}{Logical. If \code{TRUE} only the standard deviation
(a.k.a standard errors) wille be show.}

\item{...}{Not used.}
}
\description{
Extracts and show the coefficients of a \code{rqTList} object
along with their standard deviations.
}
\note{
This method is very slow, because it calls
    \code{\link[quantreg]{summary.rq}}.
}
\examples{
library(extRemes)
data(Fort)
tau <- c(0.95, 0.97, 0.98)
Rq <- list()
for (i in 1:3) {
   Rq[[i]] <- rq(Prec ~ year, data = Fort, tau = tau[i])
}
Rq <- as.rqTList(Rq)
coSd(Rq)

}
