% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exceedances.R
\name{exceed}
\alias{exceed}
\title{Exceedances Over Threshold(s)}
\usage{
exceed(object, ...)
}
\arguments{
\item{object}{An object from which exceedances can be exctracted.}

\item{...}{Arguments for further methods.}
}
\value{
Some information on the exceedances over the threshold(s).
}
\description{
Pseudo-generic (S3) function for which methods can be implemented.
}
