% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fevList.R
\name{fevdList}
\alias{fevdList}
\title{Create a \code{fevdList} Object}
\usage{
fevdList(...)
}
\arguments{
\item{...}{A collection of objects with class \code{"fevd"}
with the same \code{type}}
}
\value{
An object with class \code{"fevdList"}.
}
\description{
Create a \code{fevdList} object representing a collection of
\code{fevd} objects that can be compared e.g. graphically.
}
\details{
All the objects must have the same type and \code{"GP"} or
\code{"PP"}.
}
