% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stationsMF.R
\name{findStationMF}
\alias{findStationMF}
\title{Find a "Meteo France" Station using a Description, Id or
    Name}
\usage{
findStationMF(desc = NULL, id = NULL, name = NULL, one = TRUE)
}
\arguments{
\item{desc}{A character string to be matched against the
description of the station.}

\item{id}{A character string to be matched against the identifier
(id) of the station.}

\item{name}{A character string to be matched against the name of
the station. Character case and accents are likely to generate
problems.}

\item{one}{Logical. If \code{TRUE} the selection must correspond
to exactly one station.}
}
\value{
A one-row data frame with several elements giving the
    code, the name, the directory name for the station.  The
    column \code{Dir} gives the (a possible) name of for a
    directory containing the data.
}
\description{
Find a "Météo France" station using a description, code or name.
}
\note{
By construction, this function can only find \emph{one}
    function in order to allow automated treatments. When several
    stations match, an error with a suitable message is thrown.
}
\examples{

findStationMF(desc = "rennes-st")
findStationMF(desc = "troy")
## error
findStationMF(desc = "tro")
}
\author{
Yves Deville
}
