% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pgpTListPredSim.R
\name{format.quantile.pgpTList}
\alias{format.quantile.pgpTList}
\title{Round and Format Quantiles}
\usage{
\method{format}{quantile.pgpTList}(
  x,
  digits = 2,
  probExc = as.vector(outer(c(5, 2, 1), c(0.1, 0.01, 0.001, 1e-04))),
  ...
)
}
\arguments{
\item{x}{An object with class \code{"quantile.pgpgTList"} as
computed by the \code{quantile} method of the
\code{"pgpgTList"} class.}

\item{digits}{The number of digits to be used for the colums
\code{Quant}, \code{L} and \code{U} columns representing the
quantile and the confidence limits.}

\item{probExc}{A numeric vector giving the exceedances
probabilities to be displayed among those available in
\code{x}. The default values correspond to 12 values
\code{0.5}, \code{0.2}, \code{0.1}, ..., \code{1e-4}.}

\item{...}{Not used.}
}
\value{
A data frame with the rounded quantiles and confidence
    limits at the" chosen "round" exceedance probabilities.
}
\description{
Round the values of the quantiles and confidence limits and select
those corresponding to "round" exceedance probabilities.
}
\details{
It is often needed to show a table of quantiles corresponding to
round exceedance probabilities such as \code{0.01},
\code{0.001}. Moreover the quantiles and confidence limits as
printed by default with number of digits which may be higher than
needed. Most often, a number of digits between 0 and 2 is hight
enough in relation with the precision of the data used to compute
the quantiles.
}
