% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miscUtils.R
\name{getFun}
\alias{getFun}
\title{Find the Name of a Function within a Namespace}
\usage{
getFun(x)
}
\arguments{
\item{x}{A character such as \code{aPkg::aFun}.}
}
\value{
A reference that can be used as 1-st argument in
    \code{\link{do.call}}.
}
\description{
Find the Name of a Function within a Namespace
}
\examples{
getfun("lme4::lmer")

}
\author{
MrFlick on StackOverFlow, see
    \url{https://stackoverflow.com/a/38984214/8836534}
}
\keyword{internal}
