% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pgpTList.R
\name{isNested}
\alias{isNested}
\title{Checks that two \code{pgpTList} Objects correspond to
    Nested Models.}
\usage{
isNested(object, object1, ...)
}
\arguments{
\item{object, object1}{Two \code{pgpTlist} objects for which we
want to know if \code{object} is nested in \code{object1}.}

\item{...}{Not used}
}
\value{
Either the logical \code{TRUE} or a character message
    indicating the source from which one can conclude that the
    models are not nested.
}
\description{
Checks that two \code{pgpTList} Objects correspond to
    Nested Models.
}
