% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pgpTList.R
\name{logLik.pgpTList}
\alias{logLik.pgpTList}
\title{logLik Method for `pgpTList` Objects.}
\usage{
\method{logLik}{pgpTList}(object, ...)
}
\arguments{
\item{object}{A \code{pgpTList} object representing a list of
Poisson-GP fitted models}

\item{...}{Not used yet.}
}
\value{
A numeric matrix with two columns containing the
    log-likelihoods for the "GP" part and the "time Poisson"
    part. This matrix has as a \code{"df"} attribute a vector with
    length 2 giving the degrees of freedom used by the two parts
    of the model. While the log-likelihoods differ across
    thresholds, the degrees of freedom are the same by
    construction.
}
\description{
logLik Method for `pgpTList` Objects.
}
\section{Caution}{
 One should not compare log-likelihoods for POT
    models using different thresholds. So the comparison only
    makes sense when one use two \code{pgpTlist} objects using the
    same probabilities \code{tau} and the same formulas in the
    quantile regression as well as the same data.
}

