% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pgpTListPredSim.R
\name{makeNewData.pgpTList}
\alias{makeNewData.pgpTList}
\title{Prepare a Data Object for a Prediction}
\usage{
\method{makeNewData}{pgpTList}(object, newdata = NULL, trace = 0, ...)
}
\arguments{
\item{object}{An object with class \code{"PgpTList"}}

\item{newdata}{A "sketch of data" allowing the construction. This
can be a \code{dailyMet} object, a data frame or simply a
vector with class \code{"Date"}. With the default value, the
new data is taken as \code{object$data}.}

\item{trace}{Integer level of verbosity.}

\item{...}{Arguments for methods.}
}
\value{
A \code{dailyMet} object with the suitable variables.
}
\description{
The object given in \code{object} can be used to compute a
prediction on  "new" data.
}
