% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pgpTListPredSim.R
\name{makeNewData}
\alias{makeNewData}
\title{Prepare a Data Object for a Prediction}
\usage{
makeNewData(object, ...)
}
\arguments{
\item{object}{An object from a class having a \code{predict}
method.}

\item{...}{Arguments for methods.}
}
\value{
A "data" object that can be used as value for the
    \code{newdata} argument of the \code{predict} method.
}
\description{
When using a 'time varying' model, some functions of the
\code{Date} variable are required to compute the prediction. The
\code{makeNewData} method can be used for that goal.
}
