% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fevdMethods.R
\name{modelMatrices.fevd}
\alias{modelMatrices.fevd}
\title{Extract or Compute the Model Matrices related to a
    \code{fevd} Object.}
\usage{
\method{modelMatrices}{fevd}(object, newdata = NULL, threshold = NULL, trace = 1, ...)
}
\arguments{
\item{object}{An object with class \code{"fevd"} representing a
non-stationary POT model, either Poisson-GP (\code{type =
"GP"}) or NHPP (\code{type = "PP"}).}

\item{newdata}{A data frame containing the covariates.}

\item{threshold}{The threshold. This can be either a suitable
vector of an object representing a quantile regression,
inheriting from \code{"rq"}.}

\item{trace}{Level of verbosity}

\item{...}{Not used yet.}
}
\value{
A list of "design" matrices.
}
\description{
Extract or Compute the Model Matrices related to a
    \code{fevd} Object.
}
\examples{

## use centimetres as precipitation unit
data(Fort) 
Fort <- within(Fort, Prec <- 2.54 * Prec)
tau <- 0.97; u0 <- quantile(Fort$Prec, prob = tau)
fit0 <- fevd(x = Prec, data = Fort, threshold = u0, type = "GP")
p0 <- predict(fit0)
rq <- rq(Prec ~ cos(2 * pi * tobs / 365.25) + sin(2 * pi * tobs / 365.25),
         data = Fort, tau = tau)
summary(rq)
## compute the threshold and extract the fitted coefficients
u <- predict(rq, newdata = Fort)
beta_u <- coef(rq)
fit1 <- fevd(x = Prec, data = Fort, type = "GP",
             scale.fun = ~ cos(2 * pi * tobs / 365.25) + sin(2 * pi * tobs / 365.25),
             threshold = beta_u,
             threshold.fun = ~ cos(2 * pi * tobs / 365.25) + sin(2 * pi * tobs / 365.25))

mM <- modelMatrices(fit1)
names(mM)
head(mM$threshold)
lapply(mM, dim)

}
\author{
Yves Deville
}
