% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fevdMethods.R
\name{modelMatrices}
\alias{modelMatrices}
\title{Extract or Compute the Model Matrices related to an Object}
\usage{
modelMatrices(object, ...)
}
\arguments{
\item{object}{An object with class \code{"fevd"}.}

\item{...}{Arguments for methods.}
}
\description{
This (pseudo) S3 generic function aims at simlifying the
computations related to fitted models, such as confidence
intervals.
}
