% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miscUtils.R
\name{namesFromDates}
\alias{namesFromDates}
\title{Create Names from Dates}
\usage{
namesFromDates(date)
}
\arguments{
\item{date}{A vector that can be coerced into a \code{Date}
object.}
}
\value{
A character vector of distinct formated dates.
}
\description{
Create names from dates. Given dates that are distinct, the
function chooses the shortest format for which the formated dates
are distinct.
}
\details{
This function is useful to name the variables in a spline basis in
relation with the knots given as dates. Since these functions will
be used to describe slowly varying trends, the knots will most
often be chosen in different years and the year may be sufficient
to identify them.
}
\section{Caution}{
 in general the dates can not be retrieved from
    the names, so these should be stored to get maintain a full
    description of the basis.
}

\examples{
namesFromDates(c("1970-01-01", "1990-01-01", "1991-01-01"))
namesFromDates(c("1970-01-01", "1990-01-01", "1990-02-01"))
namesFromDates(c("1970-01-01", "1990-01-01", "1990-01-02"))
try(namesFromDates(c("1970-01-01", "1990-01-02", "1990-01-02")))

}
\keyword{internal}
