% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miscUtils.R
\name{nonNaIntervals}
\alias{nonNaIntervals}
\title{Compute the Beginning and the End of Each Non-NA Interval}
\usage{
nonNaIntervals(y)
}
\arguments{
\item{y}{A numeric vector assumed to represent the values of a
regularly sampled time series.}
}
\value{
A list of two vectors \code{start} and \code{end}.
}
\description{
Compute the begining and the end of each non-NA interval.
}
\keyword{internal}
