% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pgpTListPredSim.R
\name{parInfo}
\alias{parInfo}
\title{Provide Information about the Parameters of a Fitted Model
    Object}
\usage{
parInfo(object, ...)
}
\arguments{
\item{object}{The object for which the information will be shown.}

\item{...}{Other arguments for methods.}
}
\description{
This S3 generic function is defined in order to allow the
implementation of methods.
}
