% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pgpTList.R
\name{pgpTList}
\alias{pgpTList}
\title{Fit a non-stationary Poisson-GP Model using several
    Thresholds computed by Quantile Regression.}
\usage{
pgpTList(
  dailyMet,
  subset = NULL,
  thresholds,
  declust = TRUE,
  tauRef = 0.95,
  fitLambda = FALSE,
  logLambda.fun = ~1,
  scale.fun = ~Cst + sinjPhi1 + sinjPhi2 + sinjPhi3 - 1,
  shape.fun = ~1,
  extraDesign = NULL,
  trace = 1
)
}
\arguments{
\item{dailyMet}{An object with class \code{"dailyMet"} containing
the data.}

\item{subset}{A character defining a condition used to subset the
data, typically to select a period within the year e.g.,
summer. Note that the condition is applied \emph{after
declustering} and there are side effects: The exceedances in
the first two days or last two days of a period within year
may be lost. \emph{NOT IMPLEMENTED YET}. Note that the
condition is used only for the Extreme Value part of the
model, not on the determination of the threshold. Therefore,
\emph{the fitted thresholds remain unchanged} whether the the
condition is used or not.}

\item{thresholds}{An object with class \code{"rqTList"} containing
the thresholds.}

\item{declust}{Logical. If \code{TRUE} the exceedances over each
threshold will be declustered.}

\item{tauRef}{The reference probability defining the quantile
regression fit that will be used to define the phases of the
sine waves.}

\item{fitLambda}{Logical. If \code{TRUE} the temporal Poisson
process is fitted, using the formula \code{logLambda.fun}.}

\item{logLambda.fun}{Formula for the log-rate of the time Poisson
process. For now, there are only a few possibilities. See
\bold{Details}. Mind that \emph{only numeric covariates can be
used}.}

\item{scale.fun, shape.fun}{Formulas for the GP scale and shape as
in \code{\link[extRemes]{fevd}}.}

\item{extraDesign}{A list with a specific structure used to
generate extra "design variables" that can be used in
formulas, see \bold{Examples} and the help of
\code{\link{designVars}} and \code{\link{rqTList}} for
examples of syntax. By default a trigonometric design with
three harmonics is used, along with the sinus waves basis
corresponding to the phases of the reference threshold defined
by \code{tauRef}. The sinus wave variables are created by
using the \code{\link{sinBasis}} function with the relevant
value of \code{phi}.}

\item{trace}{Integer level of verbosity.}
}
\value{
An object with class \code{"pgpTList"}. This is a list with
    the following elements
    \describe{
        \item{\code{tau} }{

            The vector of probabilities extracted from
            \code{thresholds}.

        }
        \item{\code{thresholds} }{

            A copy of \code{thresholds} as given on entry.

        }
        \item{\code{GP} }{

            A list with class \code{\link{fevdTList}} 

        }
        \item{\code{logLambda.fun}, \code{scale.fun}, \code{scale.fun} }{

             The formulas used for the "GP part" of the model.

        }
    }
}
\description{
Fit a non-stationary Poisson-GP model for several thresholds given
in the \code{rqTList} object \code{thresholds} corresponding to a
vector of probability.
}
\details{
For each probability \code{tau[i]}, one proceeds to the following
steps

\describe{
    \item{Exceedances }{
        Find the exceedances over the threshold corresponding
        to the probability \code{tau[i]}, and decluster these
        if wanted.
    }
    \item{GP part}{
        Fit a non-stationary GP model using
        \code{extRemes::fevd} with \code{type = "GP"} and
        with the formulas prescribed.
     }
     \item{Temporal part}{
        Fit a non-stationary temporal Poisson process model using
        \code{NHPoisson::fitPP.fun}. 
     }
}

The formula given in \code{logLambda.fun} will typically involve
\code{YearNum}. Note that the constant is alaways included by
the function \code{NHPoisson::fitPP.fun} so it should be discarded
from the formula when covariates are used. For instance in order
to use the covariare \code{YearNum} we must use the formla
\code{~ YearNum -1}.
}
\note{
The use of \code{subset}, still experimental, is aimed to
    focus the estimation on a period \emph{within the year},
    typically summer or winter. This should not be used to use a
    shorter timeseries. The rate of the temporal Poisson process
    is computed on the basis of the full times series, not on the
    basis of the cumulated duration of the subset. For instance if
    \code{subset} selects 6 months in a year and if the timeseries
    duration is 60 year, then the rate is still computed on the
    basis of a 60-year duration, not on a 30-year duration.
}
\section{Caution}{
 At the time being, a \code{pgpTList} object is
    not a list of \code{pgp} objects. The output is likely to be
    re-designed, so a \code{pgpTList} object is better used via
    methods.
}

\examples{
## define the thresholds with the default seasonality, see 'rqTList'
Rq <- rqTList(dailyMet = Rennes,
              tau = c(0.94, 0.95, 0.96, 0.97, 0.98, 0.99))

## fit
Pgp1 <- pgpTList(dailyMet = Rennes, thresholds = Rq,
                 declust = TRUE,
                 fitLambda = TRUE, logLambda.fun = ~ YearNum - 1)

## try a varying GP shape 'xi'
\dontrun{
    Pgp2 <- pgpTList(dailyMet = Rennes, thresholds = Rq,
                     declust = TRUE, 
                     shape.fun = ~ Cst + sinjPhi1 + sinjPhi2 + sinjPhi3 - 1,
                     fitLambda = TRUE, logLambda.fun = ~ YearNum - 1)
}
## plot one year to 
predYear <- predict(Pgp1, last = TRUE)
gYear <- autoplot(predYear, facet = FALSE)
gYear

## show the evolution of the exceedance rate on the long-run
predAll <- predict(Pgp1, last = FALSE)
exceed <- exceed(Pgp1)
gAll <- autoplot(predAll, which = "lambda", size = 1.2) +
            geom_point(data = exceed, mapping = aes(x = Date, y = Nb)) +
                ggtitle(paste("Fitted rate 'lambda' and annual number of",
                              "declustered exceedances"))
gAll

## Add an extra design corresponding to broken line splines. This
## creates a new variable 't1_1970' that can be used in formula to
## assess a possible change in the trend at the beginning of the seventies.
Pgp3 <-
    pgpTList(dailyMet = Rennes, thresholds = Rq,
             declust = TRUE,
             fitLambda = TRUE,
             logLambda.fun = ~ YearNum + t1_1970 - 1,
             extraDesign =
                 list("breaks" = list(what = "NSGEV::breaksX",
                                      args = list(breaks = "1970-01-01"))))
p3 <- predict(Pgp3,
              newdata = data.frame(Date = seq(from = as.Date("2024-01-01"),
                                   to = as.Date("2054-01-01"),
                                   by = "day")))
autoplot(p3, facet = FALSE)

## simulate 
s3 <- simulate(Pgp3, nsim = 10,
              newdata = data.frame(Date = seq(from = as.Date("2024-01-01"),
                                   to = as.Date("2054-01-01"),
                                   by = "day")))
autoplot(s3)

}
\seealso{
\code{\link{rqTList}} for the list of \code{rq} objects
    as used in \code{thresholds}, and \code{\link{fevdTList}} for
    the list of \code{fevd} objects as returned in the \code{GP}
    element of the result.
}
