% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pgpTListPredSim.R
\name{predict.pgpTList}
\alias{predict.pgpTList}
\title{Predict a `pgpTList` Object.}
\usage{
\method{predict}{pgpTList}(
  object,
  newdata = NULL,
  lastFullYear = FALSE,
  trace = 0,
  which = c("param", "max"),
  probExcM = c(0.05, 0.01, 0.005, 0.001, 5e-04, 1e-04),
  ...
)
}
\arguments{
\item{object}{A \code{pgpTList} object representing a list of
Poisson-GP fitted models}

\item{newdata}{An optional object giving the "new" dates for which
the simulation will be done. It can simply be an object with
class \code{"Date"} or an object with class
\code{"dailyMet"}. Depending on the class of \code{newdata},
the variables required for the prediction (such as sine waves)
will be recomputed or not. When \code{newdata} is not given or
is \code{NULL}, \code{object$data} is used.}

\item{lastFullYear}{Logical, used only when \code{newdata} is not
provided or is \code{NULL}. When \code{TRUE}, only the last
full year in \code{newdata} will be used.}

\item{trace}{Integer level of verbosity.}

\item{which}{The kind of result wanted.}

\item{probExcM}{A vector of (small) probabilities of exceedance
for the maximum \eqn{M} of the marks on the prediction
period. For each probability \eqn{p}, the value of the
quantile \eqn{m} of \eqn{M} with probability \eqn{1 - p} will
be computed. Note that the quantile may be \code{NA} if the
value \eqn{F_M(m)} of distribution function of \eqn{M} is
lower than \eqn{1 -p} when \eqn{m := \textrm{max}_t\{u_t\}}{m
:= max_t _(t)} because the POT model can only provide the tail
distribution of \eqn{M}.}

\item{...}{Not used yet.}
}
\value{
An object with class \code{"predict.pgpTList"} A data
    frame in long format. Among the columns we find \code{Date},
    \code{tau} and \code{u} and the NHPP parameters \code{muStar},
    \code{sigmaStar} and \code{xiStar}. The column \code{sigma}
    contains the GP scale parameter.
}
\description{
Predict a `pgpTList` Object.
}
\note{
Remind that the NHPP parameters do not depend on the
    threshold, although their estimates obviously do. They can be
    used to assess the sensitivity too the threshold choice.
}
\section{Caution}{
 This method still may change.
}

