% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bases.R
\name{print.phasesMatrix}
\alias{print.phasesMatrix}
\title{Print a \code{phasesMatrix} object}
\usage{
\method{print}{phasesMatrix}(x, digits = 3, ...)
}
\arguments{
\item{x}{The \code{phasesMatrix} object.}

\item{digits}{Number of digits to show.}

\item{...}{Not used.}
}
\value{
Nothing.
}
\description{
Print a \code{phasesMatrix} object
}
