% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pgpTListPredSim.R
\name{quantile.pgpTList}
\alias{quantile.pgpTList}
\title{Compute Quantiles for the Maximum the Marks of a Poisson-GP
    Model}
\usage{
\method{quantile}{pgpTList}(x, newdata = NULL, prob = NULL, level = 0.95, ...)
}
\arguments{
\item{x}{An object with class \code{"predict.pgpTList"} as created
by applying the \code{predict} method on an object with class
\code{"pgpTList"}.}

\item{newdata}{A "new" data frame or \code{Date} vector used to
define the "new" period. The quantiles will be the those of
the random maximum \eqn{M} of the marks on this period.}

\item{prob}{Vector of probabilities.}

\item{level}{The confidence level.}

\item{...}{Not used yet.}
}
\value{
A data frame with columns \code{Prob} and \code{Quant}.
}
\description{
The Poisson-GP model given in \code{x} only describes the tail of
the distribution of the maximum. So when a probability is too
small the quantile may be \code{NA}.
}
\details{
The computation of the quantiles without any inference result can
be performed at a lower cost by using the results of a
\code{predict} step it these have already been computed. Of course
the "new" period will be that which was used in the \code{predict}
step and can not be changed.
}
\examples{
RqU <- rqTList(dailyMet = Rennes, tau = c(0.94, 0.95, 0.96, 0.97, 0.98, 0.99))
Pgp1 <- pgpTList(dailyMet = Rennes, thresholds = RqU, declust = TRUE,
                 fitLambda = TRUE, logLambda.fun = ~YearNum - 1)
Date <- seq(from = as.Date("2020-01-01"), to = as.Date("2050-01-01"), by = "day")
## compute the quantile for the maximum on the "new" period
qMax <- quantile(Pgp1, newdata = Date)
autoplot(qMax)


}
\seealso{
\code{\link{predict.pgpTList}}.
}
