% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pgpTListQuant.R
\name{quantMax}
\alias{quantMax}
\title{Compute Quantiles for a Maximum}
\usage{
quantMax(x, ...)
}
\arguments{
\item{x}{An object representing a fitted model from which the
distribution of the maximum (or at least its tail) can be
found.}

\item{...}{Arguments for methods.}
}
\value{
An object containing the wanted quantiles. As a rule this
    object could have a (S3) class extending the data frame class,
    and the returned object could have columns corresponding to
    the probability and the quantile, plus other columns such as
    confidence limits and possibly a confidence level.
}
\description{
The (S3) generic function \code{quantMax} evaluates the quantiles
for a maximum as is often of interest in Extreme-Value
modelling. The methods for this function should require the
definition of a period of time, a region in space or a
spatio-temporal domain on which the (random) maximum is computed
and the probability of interest, usually close to one.
}
\details{
Note that for some tasks such as building return level plots it to
have a large number of probabilities, yet one usually focus on
"pretty" or round probabilities such as \code{0.9}, \code{0.99}
and \code{0.999} to summarize the results.
}
