% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readFiles.R
\name{readECA}
\alias{readECA}
\title{Read an ECA File and Add Extra Varables}
\usage{
readECA(file, out = c("dailyMet", "data.frame"), station = NA, code = NA)
}
\arguments{
\item{file}{The file to read as in \code{\link{read.table}}.}

\item{out}{The class of the returned object. The \code{"dailyMet"}
S3 class with creator \code{\link{dailyMet}} extends the
\code{data.frame} by attaching some information to the data as
attributes. This information inclues the name of the
meteorological variable.}

\item{station, code}{Optional character vectors with length one
that will be attached to the object when the output class is
\code{"dailyMet"}.}
}
\value{
A data frame with the temperature \code{TX}, the date
    \code{Date} and several functions of the Date.

A data frame with the columns read in the file and
    possibly also those built by the \code{\link{dailyMet}}
    function.
}
\description{
Read a file as provided by the ECA Data  with \code{StatId}
}
\section{Caution}{
 For now, only ECA files for so called
    \emph{blended} data are used, and the variable is assumed to
    be a temperature. Also the \code{SOUID} field identifying the
    source is discarded.

 The station identifier \code{StatId} used by
    European Climate Assessment & Dataset project (ECA&D)
    \url{https://www.ecad.eu/} is different from the one used by
    Météo France. Also ECA&D makes a distinction between
    \emph{Station Id} (\code{StatId}) and \emph{Source Id}
    (\code{SouId}).
}

