% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readFiles.R
\name{readMet}
\alias{readMet}
\title{Read a csv File Containing Daily Meteorological Timeseries
    and Add Extra Variables}
\usage{
readMet(station, metVar = "TX", out = c("dailyMet", "data.frame"))
}
\arguments{
\item{station}{A list as returned by \code{\link{findStationMF}}.}

\item{metVar}{The meteorological variable.}

\item{out}{The class of the returned object.}
}
\value{
An object inheriting from the \code{"data.frame"} class,
    by default a \code{dailyMet} object.
}
\description{
Read a csv file containing a daily meterorological timeseries as
a column. The file must be a \code{.csv} file (\emph{Comma
Separated Values}) using the semi-column \code{";"} as column
delimiter. It is found in the directory defined by the
\code{metData} environment variable.
}
\details{
The fields are as follows.
\describe{  
   \item{\code{Code} }{
       Character: the code to be used for the station}
   \item{\code{Date} }{
       The date in POSIX format Y-m-d}
   \item{TX}{
      The temperature in Celsius.
   }
   \item{\code{Source} }{
      Character code giving the source of the data, useful if
      several sources are used in "blended" .
    }
}
}
\note{
The files that can be read are located in the repository
    defined in the system variable \code{metData}. You may need to
    set this variable by using \code{\link{Sys.setenv}}, possibly
    in a \code{Rprofile} of \code{.Renviron} file.
}
\examples{
\dontrun{
   ## read data for Bordeaux-Merignac
   s <- findStationMF("bordeaux-mer")
   Met <- readMet(s)
}

}
