% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pgpTListResid.R
\name{residuals.pgpTList}
\alias{residuals.pgpTList}
\alias{resid.pgpTList}
\title{Generalized Residuals for a \code{TVGEV} Model}
\usage{
\method{residuals}{pgpTList}(object, type = c("exp", "unif"), ...)
}
\arguments{
\item{object}{A \code{TVGEV} object.}

\item{type}{The approximate distribution wanted. The choices
\code{c("unif", "exp")} correspond to the standard uniform,
the standard exponential and the standard Gumbel
distributions. Partial matching is allowed.}

\item{...}{Not used yet.}
}
\value{
A vector of generalized residuals which should
    \emph{approximately} be independent and \emph{approximately}
    follow the standard exponential or the uniform distribution,
    depending on the value of \code{type}.
}
\description{
Generalized Residuals for a \code{TVGEV} model.
}
\examples{
if (require("NSGEV")) {
## build thresholds then fit
Rq <- rqTList(dailyMet = Rennes, tau = c(0.94, 0.95, 0.96, 0.97, 0.98, 0.99))
Pgp <- pgpTList(dailyMet = Rennes, thresholds = Rq, declust = TRUE,
                extraDesign  = list(splines = list("what" = "NSGEV::breaksX",
                                                   "args" = list(breaks = "1980-01-01"))),
                scale.fun = ~Cst + sinjPhi1 + sinjPhi2 + sinjPhi3 + t1_1980 - 1,
                fitLambda = TRUE, logLambda.fun = ~ t1_1980 - 1)

res <- resid(Pgp)
autoplot(res)
autoplot(res, seas = TRUE)
}
}
\references{
Cox, D.R. and Snell E.J. (1968) "A General Definition
of Residuals".  \emph{JRSS Ser. B}, \bold{30}(2), pp. 248-275.
}
