% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miscUtils.R
\name{seasCenter}
\alias{seasCenter}
\title{Extract a Time Window in Year from a Data Frame.}
\usage{
seasCenter(data, j, sw = 45)
}
\arguments{
\item{data}{The data frame. Must have a \code{Day} column giving
the Julian day.}

\item{j}{The center Julian day. Can be given as an integer between
1 and 365 or as character string such as \code{"07-21"} for
the 21-th of July.}

\item{sw}{Semi-width (in days) of the sliding period). So \code{sw
= 45} means that the fit is made on a period of 90 days.}
}
\value{
A subset data frame.
}
\description{
Extract a time window in year from a data frame.
}
\details{
Based on the column \code{Day} in the data frame, and the give
Julian day for the center we extract the corresponding
observations. This function is useful e.g., to investigate the
form of the seasonality on a parameter.
}
\examples{
df <- seasCenter(Rennes, j = "07-21")

}
