% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pgpTListPredAndPredict.R
\name{simulate.pgpTList}
\alias{simulate.pgpTList}
\title{Simulate a `pgpTList` Object.}
\usage{
\method{simulate}{pgpTList}(
  object,
  n = 1,
  newdata,
  lastFullYear = FALSE,
  tau = NULL,
  trace = 0,
  ...
)
}
\arguments{
\item{object}{A \code{pgpTList} object representing a list of
Poisson-GP fitted models}

\item{newdata}{\bold{Not implemented}. Now taken as
\code{object@data}.}

\item{lastFullYear}{Logical used only when \code{newdata} is not
provided. When \code{TRUE}, only the last full year in the
data used when fitting the object will be used.}

\item{trace}{Integer level of verbosity.}

\item{...}{Not used yet.}
}
\value{
An object with class \code{"summary.pgpTList"}, inheriting
    from the \code{"data.frame"} class. This is a data frame in
    long format containing the simulated events with the
    corresponding simulated values of the meteorological variable.
}
\description{
Simulate from a `pgpTList` Object. F>or each of the thresholds of
the object, a collection of \code{n} random drawings of the
exeedance events and the related values are provided.
}
\details{
The thresholds and hence the exceedance events are related to a
probability \code{tau} of non-exceedance which is quite high, say
0.95 or more. Note that even if the GP model for the excesses over
the threshold is perfectly adequate, the simulated exceedance
events come at a rate which is smaller than the real rate because
the simulation is based on \emph{declustered} exceedance
events. So for a given date, if a large number \code{n} of
simulations is used, the frequency of the exceedance will be
smaller than the real one.
}
\section{Caution}{
 this method still may change.
}

