% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bases.R
\name{sinPhases}
\alias{sinPhases}
\title{Phases of Sine Waves from the Trigonometric Coefficients}
\usage{
sinPhases(trigCoef)
}
\arguments{
\item{trigCoef}{A numeric vector with (odd) length \eqn{2K +1}
  having suitable names related to the trigonometric basis
\code{\link{tsDesign}}.}
}
\value{
A numeric vector of length \eqn{K} containing the phases
    \eqn{\phi_k}. This vector has as its attribute
    \code{"amplitude"} another numeric vector with length \eqn{K}
    containing the amplitudes \eqn{\gamma_k}.
}
\description{
Find the phases \eqn{\phi_k} and the amplitudes \eqn{\gamma_k}
for \eqn{k = 1}, ..., \eqn{K} such that
\deqn{ 
 \alpha_ 0 + \sum_{k=1}^K \alpha_k \cos\{\omega_k t\} +
   \beta_k \sin\{\omega_k t \}
 = \text{Cst} + \sum_{k=1}^K \gamma_k \sin\{ \omega_k [t - \phi_k] \}
}
where \eqn{\omega_k = 2 \pi k /365.25} and where the coefficients
\eqn{\alpha_k} and \eqn{\beta_k} are given in the \code{trigCoef}
vector.
}
\section{Caution}{
 the vector \code{trigCoef} must be \emph{named}
    with suitable element names in order to allow a reliable
    extraction of the coefficients \eqn{\alpha_k} and
    \eqn{\beta_k}. These correspond to the names
    \itemize{
       \item{"Cst" }{the constant \eqn{\alpha_0}}
       \item{"cosj1", "cosj2", ... }{coefficients for the cosine terms
           \eqn{\alpha_1}, \eqn{\alpha_2}, ..., \eqn{\alpha_K}}
       \item{"sinj1", "sinj2", ... }{coefficients for the sine terms
           \eqn{\beta_1}, \eqn{\beta_2}, ..., \eqn{\beta_K}}
    }
}

