% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dailyMet.R
\name{summary.dailyMet}
\alias{summary.dailyMet}
\title{Summary Method}
\usage{
\method{summary}{dailyMet}(object, ...)
}
\arguments{
\item{object}{An object to summarize.}

\item{...}{Not used.}
}
\value{
A summary object which is printed.
}
\description{
Summary method
}
