% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pgpTList.R
\name{summary.pgpTList}
\alias{summary.pgpTList}
\title{Summary Method for `pgpTList` Objects.}
\usage{
\method{summary}{pgpTList}(object, ...)
}
\arguments{
\item{object}{A \code{pgpTList} object representing a list of
Poisson-GP fitted models}

\item{...}{Not used yet.}
}
\value{
An object with class \code{"pgpTList"}.
}
\description{
Summary Method for `pgpTList` Objects.
}
