% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rqTList.R
\name{tau}
\alias{tau}
\alias{tau.rqTList}
\title{Probability for a Quantile Regression Object}
\usage{
tau(object, ...)

\method{tau}{rqTList}(object, ...)
}
\arguments{
\item{object}{Quantile regression object.}

\item{...}{Arguments for further methods.}
}
\value{
The vector of (non-exceedance) probabilitiies.
}
\description{
Probability for a Quantile Regression Object
}
