% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bases.R
\name{tsDesign}
\alias{tsDesign}
\title{Designs For Time Series Regression}
\usage{
tsDesign(
  dt,
  type = c("polynom", "trigo", "sinwave"),
  df = 3,
  period = "year",
  phi = NULL,
  keepTrig = FALSE,
  origin = NULL
)
}
\arguments{
\item{dt}{A \code{POSIXct} or \code{Date} object.}

\item{type}{Type of design.}

\item{df}{Number of basis functions, or "degree of freedom".}

\item{period}{Not used. Can only be one year and is used only in
the trigonometric case.}

\item{phi}{Numeric vector of phases used only when \code{type} is
\code{"sinwave"}.}

\item{keepTrig}{Logical. Used only when \code{type} is
\code{"sinwave"}. If \code{TRUE} the base trigonometric
functions \code{sinj1}, \code{sinj2}, ..., \code{cosj1},
\code{cosj2}, ... are kept in the result, along with the
functions \code{sinjPhi1}, \code{sinjPhi2}, ...}

\item{origin}{An origin for the time. Used only in the polynomial
case. The same value must be used in fits and predictions.}
}
\value{
A list with an element named \code{X} containing the
    matrix.
}
\description{
Create designs for time series regression (linear regression or
quantile regression) for meteorological time series.
}
\details{
The choice is for now between polynomial functions that can be
used to describe a trend) and trigonometric functions with period
one year (365.25 days) that can be used for the seasonality.

\describe{
 
    \item{"polynom" }{

        Can be used to describe a polynomial with degree
        \code{<= df - 1}.  The \code{df} basis functions are named
        \code{"Cst"}, then: \code{"t1"}, \code{"t2"}, ...

    }

    \item{"trigo" }{

         The basis contains \code{df = 2 * K + 1} trigonometric
         functions given in the order of increasing
         harmonics. These are: the constant function \code{"Cst"}
         and then the \code{K} couples of one cosine function and
         one sine function, with names \code{"cosj1"},
         \code{"sinj1"}, \code{"cosj2"}, \code{"sinj2"}, ...

    }

    \item{"sinwave" }{

         With \code{K = df \%/\% 2}, the \code{K + 1} basis
         functions are the constant \code{"Cst"}, then \code{K}
         sine wave functions in the order of increasing harmonics
         with names \code{"sinjPhi1"}, \code{"sinjPhi2"}, ...

    } 
}
}
\note{
For both designs \code{"trigo"} and \code{"sinwave"},
    \code{df} must be taken as an \emph{odd} integer, in order to
    allow the use of an arbitrary phase for each of the harmonics.
    Mind that the corresponding bases are (nearly) orthogonal only
    when full years are used.
}
