% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rqTList.R
\name{xi}
\alias{xi}
\title{Compute an Estimate of the Tail Coefficient 'xi'}
\usage{
xi(object, ...)
}
\arguments{
\item{object}{Object from which the estimation is to be done.}

\item{...}{Arguments for further methods.}
}
\value{
Estimation(s) of \eqn{\xi} in a vector or data frame.
}
\description{
Compute an Estimate of the Tail Coefficient 'xi'
}
