% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rqList.R
\name{xi.rqTList}
\alias{xi.rqTList}
\title{Compute an Estimate of the Tail Coefficient 'xi' using
    Quantile Regression Results}
\usage{
\method{xi}{rqTList}(object, tau = NULL, lastFullYear = TRUE, plot = FALSE)
}
\arguments{
\item{object}{An object with class \code{"rqTList"}.}

\item{tau}{A vector of 3 probabilities. This vector must be in
strictly increasing order and contain values that are found in
\code{tau(object)}. By default the three largest values in
\code{tau(object)} are used.}
}
\value{
A data frame with a \code{xiHat} column containing the
    estimation.
}
\description{
Estimate the tail coefficient \eqn{\xi} from quantiles.
}
\details{
Let \eqn{\tau_1 < \tau_2 < \tau_3} be the probabilities and
\eqn{T_1 < T_2 < T_3} be
the corresponding return periods \eqn{T_i = 1 / (1 - \tau_i)}. Let
\eqn{q_1}, \eqn{q_2}, \eqn{q_3} be the corresponding quantiles as computed
by quantile regression. Then we can find \eqn{\xi} by solving the equation
\deqn{\dfrac{T_3^\xi - T_1^\xi}{T_2^\xi - T_1^\xi} =
   \dfrac{q_3 - q_1}{q_2 - q_1}.}{[T_3^xi - T_1^xi] / [T_2^\xi - T_1^\xi] =
 [q_3 - q_1] / [q_2 - q_1].} 
The value of \eqn{\xi} then depends on the covariates used in the quantile
regression. We only consider here the case where the quantiles depend on
the date, with emphasis on the case where the quantile are periodic
functions of the date with one-year period. The the value of \eqn{\xi}
has also the one-year periodicity, and we can assess its variation
on a one-year period which can be the last full year in the period used
to estimate the quantiles.
}
\section{Caution}{
 There may be a huge uncertainty on this
    estimate, and it seems that the variations are somewhat
    exaggerated. An alternative method is estimating \eqn{xi} by
    using a moving time window in the year.
}

