% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fevdTList.r
\name{as.fevdTList}
\alias{as.fevdTList}
\title{Coerce into a \code{fevdTList} object.}
\usage{
as.fevdTList(object, names)
}
\arguments{
\item{object}{An object to coerce, typically a list object.}
}
\value{
An object with S3 class \code{"fevdTList"}
}
\description{
Coerce into a \code{fevdTList} object.
}
