% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rqTList.R
\name{as.rqTList}
\alias{as.rqTList}
\title{Create a \code{rqTList} Object}
\usage{
as.rqTList(object)
}
\arguments{
\item{object}{An object to coerce, typically a list object.}

\item{...}{A collection of objects with class \code{"rq"}
with the same \code{type}}
}
\value{
An object with class \code{"rqTList"}.

##@export

An object with S3 class \code{"rqTList"}
}
\description{
Create a \code{rqTList} object representing a collection of
\code{rq} objects that differ only by the value of \code{tau}.
This class is useful to compare the object e.g., graphically.
}
\note{
It is a good practice to use a \emph{named} list.
}
\section{Caution}{
 This class contains lists of \code{rq} objects
    sharing the \emph{same formula} and only differing by the
    value of the probability \code{tau}.
}

