% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rqList.R
\name{coSd.rqTList}
\alias{coSd.rqTList}
\title{Coefficients and Standard Errors}
\usage{
\method{coSd}{rqTList}(object, sOnly = FALSE, ...)
}
\arguments{
\item{object}{An object with class \code{"rq"}.}

\item{sOnly}{Logical. If \code{TRUE} only the standard deviation
(a.k.a standard errors) wille be show.}

\item{...}{Not used.}
}
\description{
Extracts and show the coefficients of a \code{rqTList} object
along with their standard deviations.
}
\examples{
u0 <- quantile(Fort$Prec, prob = c(0.95, 0.97, 0.98))
Fits <- list()
for (i in 1:3) {
   Fits[[i]] <- rq(x = Prec, data = Fort, threshold = u0[i])
}
class(Fits) <- "rqTList"
coSd(Fits)

}
