% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dailyMet.R
\name{dailyMet}
\alias{dailyMet}
\title{Create an Object with S3 Class \code{"dailyMet"}}
\usage{
dailyMet(
  data,
  dateVar = "Date",
  metVar = "TX",
  id = list(Code = NULL),
  trace = 0,
  ...
)
}
\arguments{
\item{data}{An object than can be coerced into a data frame with
at least: a column representing the date and a column giving
the meteorological variable.}

\item{dateVar}{Name of the date variable in \code{data}. This
variable should be coercible into a vector with class
\code{"Date"} by using \code{as.Date}. If no value is given
but only a variable with class \code{"Date"} exists, then this
variable is used and renamed into \code{"Date"}.}

\item{metVar}{Name of the meteorological variable in `data`.}

\item{...}{Not used yet}

\item{idVar}{List defining of variables that can be found of
created from those in the data.frame, and that will be kept
under the prescribed name. See \bold{Examples}. \bold{NOT
IMPLEMENTED}.}
}
\value{
An object with S3 class \code{"dailyMet"} inheriting from
    the \code{"data.frame"} class.
}
\description{
Create an object with class \code{"dailyMet"} from a data frame.
}
\details{
The \code{"dailyMet"} class essentially a daily time series with
precomputed time variables such as the Julian day, ...
\itemize{
   \item{\bold{Regular sampling} }{
      The rows must correspond to a sequence of successive days.
      So missing values will generally be present.
    }

    \item{\bold{Precomputed variables} }{ These variables are
        functions of the date that are repeatedly used, hence
        which can be computed only once. They include for instance
        the Julian day.}

}
}
