% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dailyMet.R
\name{dailyMet}
\alias{dailyMet}
\title{Create an Object with S3 Class \code{"dailyMet"}}
\usage{
dailyMet(
  data,
  dateVar = "Date",
  metVar = "TX",
  idVar = list(Id = NULL),
  station = NA,
  id = NA,
  trace = 0,
  ...
)
}
\arguments{
\item{data}{An object than can be coerced into a data frame with
at least: a column representing the date and a column giving
the meteorological variable.}

\item{dateVar}{Name of the date variable in \code{data}. This
variable should be coercible into a vector with class
\code{"Date"} by using \code{as.Date}. If no value is given
but only a variable with class \code{"Date"} exists, then this
variable is used and renamed into \code{"Date"}.}

\item{metVar}{Name of the meteorological variable in `data`.}

\item{idVar}{List defining of variables that can be found or
created from those in the data.frame, and that will be kept
under the prescribed name. \bold{NOT IMPLEMENTED}.}

\item{station, id}{Character vectors with length one giving the
name of the gauging station and the identifier. These will be
attached to the data.}

\item{...}{Not used yet}
}
\value{
An object with S3 class \code{"dailyMet"} inheriting from
    the \code{"data.frame"} class.
}
\description{
Create an object with class \code{"dailyMet"} from a data frame.
}
\details{
The \code{"dailyMet"} class essentially a daily time series with
precomputed time variables such as the Julian day, ...
\itemize{
   \item{\bold{Regular sampling} }{
      The rows must correspond to a sequence of successive days.
      So missing values will generally be present.
    }

    \item{\bold{Precomputed variables} }{ These variables are
        functions of the date that are repeatedly used, hence
        which can be computed only once. They include for instance
        the Julian day.}

}

The precomputed variables include the following ones.

\itemize{

    \item{\code{Day}, \code{DateRef} }{

         Day in the year a.k.a. Julian day, and reference Date.
         \code{Day} takes integer value between 1 and
         366. \code{DateRef} is a refrence date, corresponding to
         the day but in a non leap year taken as reference.  This
         date allows to plot the series against the day in the
          year still using months on the x axis.
    }

    \item{\code{DayW}, \code{DateRefW} }{

         Similar to \code{Day} and \code{DateRef}, but for
         "winter years" begining on the 1-st of August and ending
         on the 31-th of July. Winter years are useful when the
         interest is on event that usually happen in winter
         (extreme sea levels, extreme snowfall, ...).

    }

    \item{JJA, DJF }{

         Logical variables indicating the periods June-July-August 
         and December-January-Febuary which are sometimes used in the
         study of summer and winter extremes.

    }
      
}
}
\note{
We recommend to use the staion Ids and names given in
    \code{\link{stationsMF}}.
}
