% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exceedances.R
\name{exceed.pgpTList}
\alias{exceed.pgpTList}
\title{Exceedances Over Threshold(s)}
\usage{
\method{exceed}{pgpTList}(object, ...)
}
\arguments{
\item{object}{The object to plot.}

\item{...}{Not used.}
}
\description{
Exceedances Over Threshold(s)
}
