% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fevList.R
\name{fevdTList}
\alias{fevdTList}
\title{Create a \code{fevdTList} Object}
\usage{
fevdTList(...)
}
\arguments{
\item{...}{A collection of objects with class \code{"fevd"}
with the same \code{type}}
}
\value{
An object with class \code{"fevdTList"}.
}
\description{
Create a \code{fevdTList} object representing a collection of
\code{fevd} objects that can be compared e.g. graphically.
}
\details{
This class is intended to be used for models differing only by
their threshold.

\itemize{
   \item{ }{All the objects must have the same type and \code{"GP"} or
     \code{"PP"}.}
   \item{ }{All the objects must have the same formulas:
        \code{threshold.fun}, \code{location.fun}, \code{scale.fun}
        and \code{shape.fun}}
}
}
