% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pgpTListPredSim.R
\name{parInfo.pgpTList}
\alias{parInfo.pgpTList}
\title{Provide Information about the Parameters of a
    \code{pgpTList} Object.}
\usage{
\method{parInfo}{pgpTList}(object, ...)
}
\arguments{
\item{object}{A \code{pgpTList} object.}

\item{...}{Not used.}
}
\value{
A list with two elements
\itemize{
    \item{length }{
        A list containing the lengths of the blocks.
    }
    \item{names }{
        A list containing the names of the blocks and their
        elements.
    }
 }
}
\description{
Provide Information about the Parameters of a
    \code{pgpTList} Object.
}
