% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pgpTListPredSim.R
\name{parInfo}
\alias{parInfo}
\title{Provide Information about the Parameters of a Fitted Model
    Object}
\usage{
parInfo(object, ...)
}
\description{
Provide Information about the Parameters of a Fitted Model
    Object
}
