% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pgpTList.R
\name{pgpTList}
\alias{pgpTList}
\title{Fit a non-stationary Poisson-GP Model using several
    Thresholds computed by Quantile Regression.}
\usage{
pgpTList(
  dailyMet,
  subset = NULL,
  thresholds,
  declust = TRUE,
  tauRef = 0.95,
  fitLambda = FALSE,
  logLambda.fun = ~1,
  scale.fun = ~Cst + sinjPhi1 + sinjPhi2 + sinjPhi3 - 1,
  shape.fun = ~1,
  trace = 1
)
}
\arguments{
\item{dailyMet}{An object with class \code{"dailyMet"} containing
the data.}

\item{subset}{A condition used to subset the data, typically to
select a period within the year e.g., summer. Note that the
condition is applied \emph{after declustering} and there are
side effects: The exceedances in the first two days or last
two days of a period within year may be lost. \bold{NOT
IMPLEMENTED YET}.}

\item{thresholds}{An object with class \code{"rqTList"} containing
the thresholds.}

\item{declust}{Logical. If \code{TRUE} the exceedances over each
threshold will be declustered.}

\item{tauRef}{The reference probability defining the quantile
regression fit that will be used to define the phases of the
sine waves.}

\item{fitLambda}{Logical. If \code{TRUE} the temporal Poisson
process is fitted, using the formula \code{logLambda.fun}.}

\item{logLambda.fun}{Formula for the log-rate of the time Poisson 
process. For now, there are only a few possibilities. See
\bold{Details}. Mind that
\bold{only numeric covariates can be used}.}

\item{scale.fun, shape.fun}{Formulas for the GP scale and shape as
in \code{\link[extRemes]{fevd}}.}

\item{trace}{Integer level of verbosity.}
}
\value{
An object with class \code{"pgpTList"}. This is a list with
    the following elements
    \itemize{
        \item{\code{tau} }{

            The vector of probabilities extracted from
            \code{thresholds}.

        }
        \item{\code{thresholds} }{

            A copy of \code{thresholds} as given on entry.

        }
        \item{\code{GP} }{

            A list with class \code{\link{fevdTList}} 

        }
        \item{\code{logLambda.fun}, \code{scale.fun}, \code{scale.fun} }{

             The formulas used for the "GP part" of the model.

        }
    }
}
\description{
Fit a non-stationary Poisson-GP model for several thresholds given
in the \code{rqTList} object \code{thresholds} corresponding to a
vector of probability.
}
\details{
For each probability \code{tau[i]}
\itemize{
    \item{ }{
        Find the exceedances over the threshold corresponding
        to the probability \code{tau[i]}, and decluster these
        if wanted.
    }
    \item{ }{
        Fit a non-stationary GP model using
        \code{extRemes::fevd} with \code{type = "GP"} and
        with the formulas prescribed.
     }
     \item{ }{
        Fit a non-stationary Poisson process model using
        \code{NHPoisson::xxx} with \code{type = "GP"} 
     }
}
}
\section{Caution}{
 At the time being, a \code{pgpTList} object is
    not a list of \code{pgp} objects. The output is likely to be
    re-designed, so a \code{pgpTList} object is better used via
    methods.
}

\examples{
## define the thresholds with the default seasonality, see 'rqTList'
Rq <- rqTList(dailyMet = Rennes,
              tau = c(0.94, 0.95, 0.96, 0.97, 0.98, 0.99))

## fit
Pgp1 <- pgpTList(dailyMet = Rennes, thresholds = Rq,
                 declust = TRUE,
                 fitLambda = TRUE, logLambda.fun = ~ YearNum - 1)

## try a varying GP shape 'xi'
\dontrun{
    Pgp2 <- pgpTList(dailyMet = Rennes, thresholds = Rq,
                     declust = TRUE, 
                     shape.fun = ~ Cst + sinjPhi1 + sinjPhi2 + sinjPhi3 - 1,
                     fitLambda = TRUE, logLambda.fun = ~ YearNum - 1)
}
## plot one year to 
predYear <- predict(Pgp1, last = TRUE)
gYear <- autoplot(predYear, facet = FALSE)
gYear

## show the evolution of the exceedance rate on the lon-run
predAll <- predict(Pgp1, last = FALSE)
exceed <- exceed(Pgp1)
gAll <- autoplot(predAll, which = "lambda", size = 1.2) +
            geom_point(data = exceed, mapping = aes(x = Date, y = Nb)) +
                ggtitle(paste("Fitted rate 'lambda' and annual number of",
                              "declustered exceedances"))
gAll

}
\seealso{
\code{\link{rqTList}} for the list of \code{rq} objects
    as used in \code{thresholds}, and \code{\link{fevdTList}} for
    the list of \code{fevd} objects as returned in the \code{GP}
    element of the result.
}
