% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pgpTList.R
\name{predict.pgpTList}
\alias{predict.pgpTList}
\title{Predict a `pgpTList` Object.}
\usage{
\method{predict}{pgpTList}(object, newdata, lastFullYear = FALSE, ...)
}
\arguments{
\item{object}{A \code{pgpTList} object representing a list of
Poisson-GP fitted models}

\item{newdata}{\bold{Not implemented}. Nowe taken as
\code{object@data}.}

\item{lastFullYear}{Logical. When \code{TRUE}, only the last full
year in \code{newdata} will be used.}

\item{...}{Not used yet.}
}
\value{
An object with class \code{"predict.pgpTList"} A data
    frame in long format. Among the columns we find \code{Date},
    \code{tau} and \code{u} and the NHPP parameters \code{muStar},
    \code{sigmaStar} and \code{xiStar}.
}
\description{
Predict a `pgpTList` Object.
}
\note{
Remind that the NHPP parameters do not depend on the
    threshold, although their estimates obviously do. They can be
    used to assess the sensitivity too the threshold choice.
}
\section{Caution}{
 this method still may change.
}

