% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fevdMethods.R
\name{quantile.fevd}
\alias{quantile.fevd}
\title{Compute Quantile for \code{fevd} Objects}
\usage{
\method{quantile}{fevd}(x, probs = c(0.9, 0.95, 0.99), data = NULL, ...)
}
\arguments{
\item{x}{An object with class \code{"fevd"}.}

\item{probs}{A vector of probabilities.}

\item{data}{A data frame containing the covariates needed.}

\item{...}{Not used yet.}
}
\value{
An object inheriting from \code{"matrix"}.
}
\description{
Compute the quantile for Non Stationary Extreme-Value
models fitted by using \code{extRemes::fevd}.
}
