% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readFiles.R
\name{readECA}
\alias{readECA}
\title{Read an ECA File and Add Extra Varables}
\usage{
readECA(file)
}
\arguments{
\item{file}{The file to read as in \code{\link{read.table}}.}
}
\value{
A data frame with the temperature \code{Temp}, the date
    \code{Date} and several functions of the Date.

A data frame with the columns
    \itemize{
        \item{Date, DateRef }{
           the date and the "refrence" date (corresponding to
           the same month and day) in a reference leap year chosen
           as 1972, since 1970 was not a leap year.
         }
         \item{Temp }{
           the temperature.
         }
         \item{Dec, YearDec }{

            factors related to the decade as required in
            plots. \code{Dec} is the decade, and \code{YearDec} is
            the year in the decade with values from \code{"0"} to
            \code{"9"}.

         }
         \item{YearW, DayW }{
            "winter" year and day. The winter year \code{YearW} is
             a character variable with values such as
             \code{"2022-23"}. The winter day is an integer
             between 1 (for the 1-st of August) and 366 (for the
             31-th of July in a leap year).
         }
         \item{JJA, DJF }{
             logical variables for the periods in year
             \emph{June-July-August} and
             \emph{December-January-February}. }
    }
}
\description{
Read a file as provided by the ECA Data  with \code{StatId}
}
\section{Caution}{
 For now, only ECA files for so called
    \emph{blended} data are used, and the variable is assumed to
    be a temperature. Also the \code{SOUID} field identifying the
    source is discarded.
}

