% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rqTList.R
\name{rqTList}
\alias{rqTList}
\title{Create A \code{rqTList} Object by Repeated Calls to
    \code{rq}}
\usage{
rqTList(
  formula = TX ~ Cst + cosj1 + sinj1 + cosj2 + sinj2 + cosj3 + sinj3 - 1,
  dailyMet,
  tau = c(0.5, 0.7, 0.8, 0.9, 0.95, 0.97, 0.98, 0.99),
  design = list(trigo = list(type = "trigo", df = 7))
)
}
\arguments{
\item{formula}{The formula that will be used in
\code{\link[quantreg]{rq}}.}

\item{dailyMet}{An object with class \code{"dailyMet"} that
contains the variables that will be used in the formula or in
the design function specified \code{design}.}

\item{tau}{A vector of probabilities.}

\item{design}{A list of arguments to be passed to
\code{\link{tsDesign}}. For each element of this list, the
variables of the design are added to the data frame given in
\code{dailyMet} before fiiting the model. The default creates
seven trigonometric basis functions for the first three
harmonics of the yearly seasonality.}
}
\value{
An object with class \code{"rqTList"}.
}
\description{
Create a \code{qrList} object by calling \code{\link{rq}} for each
of the probabilities given in \code{tau}, using the same formula
and data for all fits.
}
\examples{
Rq <- rqTList(dailyMet = Rennes)
coef(Rq)
autoplot(Rq)
}
