% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rqTList.R
\name{rqTList}
\alias{rqTList}
\title{Create A \code{rqTList} Object by Repeated Calls to
    \code{rq}}
\usage{
rqTList(
  formula = TX ~ Cst + cosj1 + sinj1 + cosj2 + sinj2 + cosj3 + sinj3 - 1,
  dailyMet,
  tau = c(0.5, 0.7, 0.8, 0.9, 0.95, 0.97, 0.98, 0.99),
  design = list(trigo = list(what = "tsDesign", args = list(type = "trigo", df = 7))),
  trace = 0
)
}
\arguments{
\item{formula}{The formula that will be used in
\code{\link[quantreg]{rq}}.}

\item{dailyMet}{An object with class \code{"dailyMet"} that
contains the variables that will be used in the formula or in
the design function specified \code{design}.}

\item{tau}{A vector of probabilities.}

\item{design}{A list to be passed to the \code{\link{designVars}}
function. Each element of this list is itself a named list
with an element \code{what} describing the "design function"
to call, and an argument \code{args} describing the arguments
to be passed to the design function, except its first
argument. For each element of the \code{design} list, a design
matrix \code{X} is obtained by using \code{\link{do.call}} and
the columns of \code{X} are added to the data frame given in
\code{dailyMet} before fiiting the model. The default creates
seven trigonometric basis functions for the first three
harmonics of the yearly seasonality.}

\item{trace}{Integer level of verbosity.}
}
\value{
An object with class \code{"rqTList"}.
}
\description{
Create a \code{qrList} object by calling \code{\link{rq}} for each
of the probabilities given in \code{tau}, using the same formula
and data for all fits.
}
\details{
When \code{extradesign} is used, any design function used should
fulfill the following requirements.

\enumerate{
    \item The first argument must be the date. The name of this
      argument will not be used in the call.
    \item  The returned value should be an object inheriting from
       the \code{"matrix"} class with suitable column names. It can also
       be a list containing the design
       matrix as its element named \code{"X"}.
}
The call to a design function should mention \code{Date} as the
first argument. This call refers to the \code{dailyMet} object
used, from which the \code{Date} column will be used.
}
\examples{
Rq <- rqTList(dailyMet = Rennes)
coef(Rq)
autoplot(Rq)

if (require("NSGEV")) {
 Rq1 <-
   rqTList(formula = TX ~ Cst + cosj1 + sinj1 + cosj2 + sinj2 + t1_1970 - 1,
           dailyMet = Rennes,
           design = list("trigo" = list(what = "tsDesign",
                                        args = list(type = "trigo", df = 7)),
                         "breaks"= list(what = "NSGEV::breaksX",
                                        args = list(breaks = c('1970-01-01', '1990-01-01')))))
 p1 <- predict(Rq1,
               newdata = data.frame(Date = seq(from = as.Date("2024-01-01"),
                                    to = as.Date("2054-01-01"),
                                    by = "day")))
 autoplot(p1)
}

}
