% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bases.R
\name{sinBasis}
\alias{sinBasis}
\title{Create a Basis of Sine Waves with Given Phases}
\usage{
sinBasis(dt, df = 7, phi, cst = TRUE, keepTrig = FALSE)
}
\arguments{
\item{dt}{A vector with class \code{"Date"} or \code{"POSIXct"}
representing the days at which the sine waves basis functions
are to be evaluated.}

\item{df}{Odd number as in \code{\link{tsDesign}}.}

\item{phi}{Vector of \eqn{K} phases where \eqn{K} is equal to
\code{(df - 1) / 2}.}

\item{cst}{Logical. If \code{TRUE} a constant column with value
1.0 is added as a first column.}

\item{keepTrig}{Logical. If \code{TRUE}, the cos-sin functions are
joined as columns of the result. So the returned matrix then
does not correspond to a basis. Used mainly for tests.}
}
\value{
A numeric matrix containing as columns the \eqn{K} basis
    functions \eqn{s_k(t)}, and with its rows corresponding to the
    time \eqn{t} given by \code{Date}. The names of the \eqn{K}
    sine wave function have the prefix \code{"sinjPhi"} so are
    \code{"sinjPhi1"}, \code{"sinPhij2"}, ... This is aimed to
    recall that the phases \eqn{\phi_k} were given. If \code{cst}
    is \code{TRUE} and \code{keepTrig} a constant column with
    value 1.0 is added, and if \code{keepTrig} is \code{TRUE} the
    trigonometric functions (with phase zero) are added as well.
}
\description{
Basis of sine wave functions \eqn{s_k(t)} with the phases
\eqn{\phi_k} given in \code{phi} corresponding to the periods
\eqn{365.25 / k}.
}
\details{
For a time \eqn{t} the value of the \eqn{k}-th function is given
by

 \deqn{s_k(t) := \sin\{2 \pi k [j_t - \phi_k] / 365.25)}{
   s_k(t) := sin(2 * pi * k * [j_t - \phi_k] / 365.25)}

for \eqn{k=1} to {K}, where \eqn{j_t} is the Julian day
corresponding to \eqn{t}. This function is useful to provide a
basis of functions with yearly seasonality that are suitable for
meteorological variables. For instance the phase for the first
harmonic of the daily temperature in France is such that the
annual maximal temperature "normally" happen at the end of July.
}
\note{
This functions uses the \code{\link{tsDesign}} function with
    \code{type = "trigo"}.
}
\section{Caution}{
 This function is likely to be removed. It is
    safer to use \code{\link{tsDesign}} with the argument
    \code{phi}.
}

\examples{
## Use only full years to maintain near orthogonality
df <- subset(Rennes, Year <= 2020 & Year >= 1960)
K <- 3
## Design of trigonometric functions
desTrig <- tsDesign(dt = df$Date, df = 2 * K + 1, type = "trigo")
fit <- lm(formula = TX ~ Cst + cosj1 + sinj1 + cosj2 + sinj2 + cosj3 + sinj3 - 1,
          data = data.frame(df, desTrig$X))
betaHat <- coef(fit)

## find the phases
phiHat <- phases(betaHat)
gamma <- attr(phiHat, "amplitude")

## Design of sine waves with prescribed phases
desSin <- tsDesign(dt = df$Date, df = 2 * K + 1,
                   type = "sinwave", phi = phiHat)
fit2 <- lm(formula = TX ~ Cst + sinjPhi1 + sinjPhi2 + sinjPhi3 -1,
           data = data.frame(df, desSin$X))
## should be zero
max(abs(predict(fit) - predict(fit2)))

## the amplitude and the coef
rbind(coef(fit2), c(NA, gamma))
j <- 3
indTrig <- paste0(c("cosj", "sinj"), j)
indSin <- paste0("sinjPhi", j)
plot(desTrig$X[1:366 , indTrig] \%*\% betaHat[indTrig], type = "l",
     col = "SpringGreen3", xlab = "", ylab = paste("harmonic", j))
lines(desSin$X[1:366, indSin] * gamma[j], col = "red", lwd = 2, lty = "dashed")

}
