% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stationsMF.R
\docType{data}
\name{stationsMF}
\alias{stationsMF}
\title{Stations Météo-France}
\format{
An object of class \code{data.frame} with 83 rows and 12 columns.
}
\usage{
stationsMF
}
\description{
A data frame with the following columns
\itemize{
     \item{Code, Name }{Code and name.}
     \item{Lat, Lon, Alt }{

        Latitude, Longitude (decimal degrees) and Altitude
        (metres)

      }
      \item{LongName }{

         The long "exact" name in correct French. Mind spaces,
         accents and other special characters. Also compound names
         such can be a combination of two names, which is classical
         for airports.

      }
      \item{Clim }{A factor giving the climatic class, with the followin codes. 
          \itemize{
             \item{\code{"C"} }{\emph{Modified Oceanic}, mainly for Central zone
               located at the North of Clermont-Ferrand.}
             \item{\code{"Med"} }{\emph{Mediterranean}. This coevers the whole
                   French Mediterranean coastline and at the South of Valence,
                   including Toulouse.}
             \item{\code{"NW"} }{\emph{North-West Oceanic}, mainy for
                North-West. This covers the Atlantic coastline from les Sables
                d'Olonne (South) to Dunkerque (North) and covers
                French Brittany, Normandy and Picardie.}
             \item{\code{"SW"} }{\emph{Aquitain Oceanic}, mainy for South-West.
                This covers the Atlantic coasline from les Sables d'Olonnes (North)
                to Hendaye (South).}
             \item{\code{"NE"} }{\emph{Semi-Continental}, mainy for North-East.}
             \item{\code{"Mont"} }{\emph{Montaneaous}. The Massif Central and
                the French Alpes, and the French part of the Pyrenees.}
             \item{\code{""} }{Special: Tropical, Artic, ...} 
          }
      }

}
}
\keyword{datasets}
